+++ plugins/spreadsheet-view/src/SpreadsheetView/components/Spreadsheet.test.ts
@@ -7,7 +7,7 @@ describe('num to col name', () => {
     [0, 'A'],
     [25, 'Z'],
     [27 * 26 - 1, 'ZZ'],
-  ]
+  ] as const
 
   cases.forEach(([input, output]) => {
     test(`${input} -> ${output}`, () => {
